Move Operations class to separate file
